requirements.txt
psycopg2-binary
pika
redis
python-dotenv
